requirements.txt
promise
cryptography
fabric
python-dotenv